+++ src/main/java/de/cytooxien/realms/api/event/RealmPlayerJoinEvent.java
@@ -1,26 +1,36 @@
 package de.cytooxien.realms.api.event;
 
 import de.cytooxien.realms.api.enums.JoinReason;
-import net.kyori.adventure.text.Component;
 import org.bukkit.entity.Player;
-import org.bukkit.event.player.PlayerJoinEvent;
+import org.bukkit.event.HandlerList;
+import org.bukkit.event.player.PlayerEvent;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 /**
  * Event triggered when a player joins a realm.
- * Contains the reason for joining as well as an optional join message.
+ * Contains the reason for joining.
  */
-public class RealmPlayerJoinEvent extends PlayerJoinEvent {
+public class RealmPlayerJoinEvent extends PlayerEvent {
+
+    private static final HandlerList HANDLER_LIST = new HandlerList();
 
     private final JoinReason joinReason;
 
-    public RealmPlayerJoinEvent(@NotNull Player playerJoined, @Nullable Component joinMessage, @NotNull JoinReason joinReason) {
-        super(playerJoined, joinMessage);
+    public RealmPlayerJoinEvent(@NotNull Player playerJoined, @NotNull JoinReason joinReason) {
+        super(playerJoined);
         this.joinReason = joinReason;
     }
 
     public JoinReason getJoinReason() {
         return this.joinReason;
+    }
+
+    public static @NotNull HandlerList getHandlerList() {
+        return HANDLER_LIST;
+    }
+
+    @Override
+    public @NotNull HandlerList getHandlers() {
+        return HANDLER_LIST;
     }
 }